[bugfix] 修复所有demo的编译错误和警告

+++ ListDemo/MenuWnd.h
@@ -59,7 +59,7 @@ public:
 
         BLENDFUNCTION bf = { AC_SRC_OVER, 0, 255, AC_SRC_ALPHA };
         POINT ptPos   = { rcWnd.left, rcWnd.top };
-        SIZE sizeWnd  = { dwWidth, dwHeight };
+        SIZE sizeWnd  = { static_cast<LONG>(dwWidth), static_cast<LONG>(dwHeight) };
         POINT ptSrc   = { 0, 0 };
         UpdateLayeredWindow(m_hWnd, hDcPaint, &ptPos, &sizeWnd, hDcBackground, &ptSrc, 0, &bf, ULW_ALPHA);
